server/requirements.txt
requests
flask
flask-cors
psycopg2-binary
